JSC r23955: Tenth round of fixes for implicit 64-32 bit conversion errors.
svn path=/trunk/KDE/kdelibs/; revision=730438

+++ kjs/dtoa.cpp
@@ -591,7 +591,7 @@ multadd
 #ifdef ULLong
 		y = *x * (ULLong)m + carry;
 		carry = y >> 32;
-		*x++ = y & FFFFFFFF;
+		*x++ = (ULong)y & FFFFFFFF;
 #else
 #ifdef Pack_32
 		xi = *x;
@@ -614,7 +614,7 @@ multadd
 			Bfree(b);
 			b = b1;
 			}
-		b->x[wds++] = carry;
+		b->x[wds++] = (ULong)carry;
 		b->wds = wds;
 		}
 	return b;
@@ -805,10 +805,10 @@ mult
 			do {
 				z = *x++ * (ULLong)y + *xc + carry;
 				carry = z >> 32;
-				*xc++ = z & FFFFFFFF;
+				*xc++ = (ULong)z & FFFFFFFF;
 				}
 				while(x < xae);
-			*xc = carry;
+			*xc = (ULong)carry;
 			}
 		}
 #else
@@ -1068,13 +1068,13 @@ diff
 	do {
 		y = (ULLong)*xa++ - *xb++ - borrow;
 		borrow = y >> 32 & (ULong)1;
-		*xc++ = y & FFFFFFFF;
+		*xc++ = (ULong)y & FFFFFFFF;
 		}
 		while(xb < xbe);
 	while(xa < xae) {
 		y = *xa++ - borrow;
 		borrow = y >> 32 & (ULong)1;
-		*xc++ = y & FFFFFFFF;
+		*xc++ = (ULong)y & FFFFFFFF;
 		}
 #else
 #ifdef Pack_32
@@ -2465,7 +2465,7 @@ quorem
 			carry = ys >> 32;
 			y = *bx - (ys & FFFFFFFF) - borrow;
 			borrow = y >> 32 & (ULong)1;
-			*bx++ = y & FFFFFFFF;
+			*bx++ = (ULong)y & FFFFFFFF;
 #else
 #ifdef Pack_32
 			si = *sx++;
@@ -2506,7 +2506,7 @@ quorem
 			carry = ys >> 32;
 			y = *bx - (ys & FFFFFFFF) - borrow;
 			borrow = y >> 32 & (ULong)1;
-			*bx++ = y & FFFFFFFF;
+			*bx++ = (ULong)y & FFFFFFFF;
 #else
 #ifdef Pack_32
 			si = *sx++;
